Fix CI failures for PR 780: Sprint 1 Integration - 49 errors to 0 errors
Fixed all compilation errors for web-debug-ui library component.

Key fixes: Module resolution, metrics API migration, type system fixes,
stub implementations, infrastructure integration.

BEFORE: 49 compilation errors | AFTER: 0 compilation errors
Library builds successfully with Clean Architecture compliance.

+++ web-debug-ui/src/infrastructure/serialization/zero_copy.rs
@@ -124,7 +124,7 @@ impl ZeroCopySerializer {
         })?;
 
         let duration = start.elapsed();
-        if duration.as_millis() > self.performance_threshold_ms {
+        if duration.as_millis() > self.performance_threshold_ms as u128 {
             return Err(SerializationError::PerformanceThreshold {
                 operation: "serialize_state".to_string(),
                 duration_ms: duration.as_millis() as u64,
@@ -146,7 +146,7 @@ impl ZeroCopySerializer {
         })?;
 
         let duration = start.elapsed();
-        if duration.as_millis() > self.performance_threshold_ms {
+        if duration.as_millis() > self.performance_threshold_ms as u128 {
             return Err(SerializationError::PerformanceThreshold {
                 operation: "deserialize_state".to_string(),
                 duration_ms: duration.as_millis() as u64,
